repo/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rop-backend"
version = "0.1.0"
description = "Restaurant Operating Platform backend foundation"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.1.0",
  "alembic>=1.13.0",
  "redis>=5.0.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-redis>=4.6.0.20241004",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_configs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
